cloud_functions/generative_ai/requirements.txt
# Function dependencies, for example:
# package>=version 
google-cloud-documentai>=0.3.0
google-cloud-storage>=1.30.0
google-cloud-bigquery>=1.26.1
google-cloud-pubsub>=1.7.0
openai==0.11.2
prettytable==2.2.0